feature/98 refactor: config 파일들의 패키지 경로 변경
