+++ src/main/java/com/grasshouse/dorandoran/config/WebMvcConfig.java
@@ -1,22 +1,32 @@
-package com.grasshouse.dorandoran.common.config;
+package com.grasshouse.dorandoran.config;
 
 import com.grasshouse.dorandoran.post.controller.PostFilterArgumentResolver;
 import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
+@RequiredArgsConstructor
 @Configuration
-public class WebConfig implements WebMvcConfigurer {
+public class WebMvcConfig implements WebMvcConfigurer {
 
     private final PostFilterArgumentResolver PostFilterArgumentResolver;
 
-    public WebConfig(PostFilterArgumentResolver PostFilterArgumentResolver) {
-        this.PostFilterArgumentResolver = PostFilterArgumentResolver;
-    }
+    @Value("${cors.url}")
+    private String url;
 
     @Override
     public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
         resolvers.add(PostFilterArgumentResolver);
+    }
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+            .allowedOrigins(url)
+            .allowedMethods("*");
     }
 }
