--- src/main/java/com/grasshouse/dorandoran/config/cors/WebMvcConfig.java
@@ -1,20 +1,0 @@
-package com.grasshouse.dorandoran.config.cors;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebMvcConfig implements WebMvcConfigurer {
-
-    @Value("${cors.url}")
-    private String url;
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**")
-            .allowedOrigins(url)
-            .allowedMethods("*");
-    }
-}
